Move notification area code to separate source files
