+++ foo_ui_columns/mw_wnd_proc.cpp
@@ -5,6 +5,7 @@
 #include "get_msg_hook.h"
 #include "rebar.h"
 #include "status_bar.h"
+#include "notification_area.h"
 
 INT_PTR g_taskbar_bitmaps[] = {IDI_STOP, IDI_PREV, IDI_PAUSE, IDI_PLAY, IDI_NEXT, IDI_RAND};
